requirements.txt
fastapi
uvicorn
etcd3
python-dotenv
protobuf